RedGhostMovementAndRespawn
Hi guys so i did the movement for the red ghost but im really struggling with the respawn and cant move on. Ive narrowed it down to an issue with the movement controller not picking up the start node when the ghost goes through it after the start but i am not sure how to fix it because my code seems fine.